___venezia/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
mercadopago==2.2.0
requests==2.31.0
Flask-Login==0.6.2
Werkzeug==2.3.7
